feat(docker): remove support to build local sources
The `publish_image` job on gitlab does not use the build cache,
probably due to the use of the Docker instruction `COPY .`.
Thus, the complete build is running twice,
and uploading to docker hub finally fails with a GPG error
(which might be caused by a timeout in the gpg-agent).
This must be investigated and fixed.